done, settings need to be saved. (feature to add) and redesign

+++ Settings.cs
@@ -11,9 +11,9 @@ namespace CardMonthlyPay
             NumMaxMoney.ValueChanged += new EventHandler(MainForm.NumMaxMoney_ValueChanged);
             NumPayDay.ValueChanged += new EventHandler(MainForm.NumPayDay_ValueChanged);
         }
-        public Settings(int NumMaxMoney, int NumPayDay)  : this()
+        public Settings(double NumMaxMoney, int NumPayDay)  : this()
         {
-            this.NumMaxMoney.Value = NumMaxMoney;
+            this.NumMaxMoney.Value = (decimal)NumMaxMoney;
             this.NumPayDay.Value = NumPayDay;
         }
